Django2YingOps/Django2YingOps/urls.py
"""Django2YingOps URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/2.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import path,include,re_path
from django.views.static import serve
from Django2YingOps.settings import MEDIA_ROOT

import xadmin
from users.views import *
from assets.views import *

urlpatterns = [
    # path('admin/', admin.site.urls),
    path('xadmin/', xadmin.site.urls),
    # include urls
    path('users/',include('users.urls')),
    path('assets/',include('assets.urls')),

    path('',LoginView.as_view(),name='login'),
    path('logout/',LogoutView.as_view(),name='logout'),
    path('index/',IndexView.as_view(),name='index'),
    path('register/',RegisterView.as_view(),name='register'),
    path('forget/',ForgetPwdView.as_view(),name='forget'),

    #配置上传文件的访问处理函数
    re_path('media/(?P<path>.*)$', serve, {"document_root":MEDIA_ROOT}),
]
