Django2YingOps/requirements.txt
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.8
django-crispy-forms==1.9.0
django-formtools==2.2
django-import-export==2.0.2
django-reversion==3.0.7
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.17.2
idna==2.9
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==1.4.6
odfpy==1.4.1
openpyxl==3.0.3
paramiko==2.7.1
Pillow==7.0.0
ply==3.11
pyasn1==0.4.8
pycparser==2.20
pycryptodomex==3.9.7
PyMySQL==0.9.3
PyNaCl==1.3.0
pysmi==0.3.4
pysnmp==4.4.12
python-nmap==0.6.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
tornado==6.0.4
urllib3==1.25.9
webssh==1.5.2
xlrd==1.2.0
XlsxWriter==1.2.8
xlwt==1.3.0
